Basic implementation of fizz buzz
Implementing the basic fizz buzz algorithm in x64 nasm assembly. The
target number is hardcoded for now to be 20. For simplicity the
actual printing of the number (if not fizz / buzz), is done via printf,
so we are linking agains the c library here. The rest is independend, as
we are using syscalls.